fix: Implemented API auth and refactored recources component
