+++ GR.Laromedel.AccessPortal.Api/Configuration/ApiConfiguration.cs
@@ -4,5 +4,6 @@
     {
         public string Authority { get; set; }
         public string Audience { get; set; }
+        public string SPAClient { get; set; }
     }
 }
