+++ GR.Laromedel.AccessPortal.Api/appsettings.json
@@ -9,6 +9,7 @@
   "AllowedHosts": "*",
   "ApiConfiguration": {
     "Authority": "https://gridpstsidentity20201020044704.azurewebsites.net",
-    "Audience": "accessPortalApi"
+    "Audience": "accessPortalApi",
+    "SPAClient": "http://localhost:1234"
   } 
 }
