+++ GR.Laromedel.AccessPortal.Web/src/config/oidcconfig.js
@@ -2,7 +2,7 @@ export const oidcConfig = {
     autoSignIn: true,
     authority: 'https://gridpstsidentity20201020044704.azurewebsites.net',
     clientId: 'accessPortalClientLocalHost',
-    redirectUri: 'http://localhost:1234',
+    redirectUri: 'http://localhost:1234/signin-callback',
     responseType: 'code',
     scope: 'openid profile roles email accessPortalApi'    
 };
